requirements.txt
pytorch>=1.4
torchvision>=0.5
opencv-python
visdom
tqdm
ipdb
fire